Separate chip.h from chip.cpp
